Viz-Katalog V1.3/V1.4: extract SVG diagram renderers + hoist viewport store
Behaviour-preserving refactor of the 2216-line DiagramView.svelte monolith
(P0 foundation for the Visualization-Catalog epic, #61). No visual changes,
no new features — pure structural split so the still-open catalog features
(Mini-Map #66, Timeline River #63.4, Snapshots #125, ...) have a testable
renderer layer + a shared viewport to build on.

Renderer extraction (all pure `render(payload, opts): string`, vitest-able,
mirroring the existing treemap.ts pattern):
- app/src/lib/diagrams/common.ts       — esc / formatBytes / shortLabel
- app/src/lib/diagrams/folderMap.ts     — solar / hierarchy / top-down layouts
                                          (colour-by fill passed in as a
                                          FillFor callback; git state stays in
                                          the component)
- app/src/lib/diagrams/docGraph.ts      — network / radial / orphans layouts
- app/src/lib/diagrams/languageStats.ts
- app/src/lib/diagrams/architectureFlow.ts
- app/src/lib/diagrams/moduleChord.ts
- app/src/lib/diagrams/activityHeatmap.ts

Viewport store (V1.4, Mini-Map prerequisite):
- app/src/lib/diagrams/viewport.ts — pan/zoom/reset as pure reducers behind a
  small per-instance Svelte store. DiagramView drives it from wheel/drag/
  toolbar; the Mini-Map (#66) can later read the viewport rect + dispatch the
  same reducers. zoomAround() keeps the anchor world-point fixed and clamps to
  [MIN_SCALE, MAX_SCALE].

DiagramView.svelte shrinks 2216 → 1243 lines: loads data, builds the fill
resolver, picks a renderer module, renders — plus the DOM-bound bits (click
drilldown, pan/zoom handlers, base-size stamping, toolbar).

Tests (vitest, all green): viewport reducer suite (pan/zoom/reset/anchor-pin/
clamp) + layout/escaping/empty-state suites for folderMap, docGraph and the
four chart renderers. Existing diagram + i18nDiagramKeys guards stay green.

Frontend-only; no Rust/Tauri crate touched.
